Coerce Trial Design Set to a Data Frame

Returns the TrialDesignSet as data frame.

Usage

# S3 method for TrialDesignSet
as.data.frame(
  x,
  row.names = NULL,
  optional = FALSE,
  niceColumnNamesEnabled = FALSE,
  includeAllParameters = FALSE,
  addPowerAndAverageSampleNumber = FALSE,
  theta = seq(-1, 1, 0.02),
  nMax = NA_integer_,
  ...
)

Arguments

x

A TrialDesignSet object.

niceColumnNamesEnabled

Logical. If TRUE, nice looking column names will be used; syntactic names (variable names) otherwise (see make.names).

includeAllParameters

Logical. If TRUE, all available parameters will be included in the data frame; a meaningful parameter selection otherwise, default is FALSE.

addPowerAndAverageSampleNumber

If TRUE, power and average sample size will be added to data frame, default is FALSE.

theta

A vector of standardized effect sizes (theta values), default is a sequence from -1 to 1.

nMax

The maximum sample size. Must be a positive integer of length 1.

...

Ensures that all arguments (starting from the "...") are to be named and that a warning will be displayed if unknown arguments are passed.

Value

Returns a data.frame.

Details

Coerces the design set to a data frame.

Examples

designSet <- getDesignSet(design = getDesignGroupSequential(), alpha = c(0.01, 0.05))
as.data.frame(designSet)
#>   designNumber typeOfDesign kMax stages informationRates alpha beta
#> 1            1           OF    3      1        0.3333333 0.025  0.2
#> 2            1           OF    3      2        0.6666667 0.025  0.2
#> 3            1           OF    3      3        1.0000000 0.025  0.2
#> 4            2           OF    3      1        0.3333333 0.010  0.2
#> 5            2           OF    3      2        0.6666667 0.010  0.2
#> 6            2           OF    3      3        1.0000000 0.010  0.2
#> 7            3           OF    3      1        0.3333333 0.050  0.2
#> 8            3           OF    3      2        0.6666667 0.050  0.2
#> 9            3           OF    3      3        1.0000000 0.050  0.2
#>   twoSidedPower sided tolerance   alphaSpent criticalValues  stageLevels
#> 1         FALSE     1     1e-08 2.591737e-04       3.471091 2.591737e-04
#> 2         FALSE     1     1e-08 7.160059e-03       2.454432 7.055362e-03
#> 3         FALSE     1     1e-08 2.499999e-02       2.004036 2.253312e-02
#> 4         FALSE     1     1e-08 2.285478e-05       4.076546 2.285478e-05
#> 5         FALSE     1     1e-08 1.981813e-03       2.882554 1.972331e-03
#> 6         FALSE     1     1e-08 9.999990e-03       2.353595 9.296426e-03
#> 7         FALSE     1     1e-08 1.532590e-03       2.961125 1.532590e-03
#> 8         FALSE     1     1e-08 1.873932e-02       2.093831 1.813751e-02
#> 9         FALSE     1     1e-08 4.999999e-02       1.709606 4.366937e-02