Get Simulation Data

Returns the aggregated simulation data.

Usage

getData(x)

getData.SimulationResults(x)

Arguments

x

A SimulationResults object created by getSimulationMeans(),
getSimulationRates(), getSimulationSurvival(), getSimulationMultiArmMeans(),
getSimulationMultiArmRates(), or getSimulationMultiArmSurvival().

Value

Returns a data.frame.

Details

This function can be used to get the aggregated simulated data from an simulation results object, for example, obtained by getSimulationSurvival(). In this case, the data frame contains the following columns:

iterationNumber: The number of the simulation iteration.

stageNumber: The stage.

pi1: The assumed or derived event rate in the treatment group.

pi2: The assumed or derived event rate in the control group.

hazardRatio: The hazard ratio under consideration (if available).

analysisTime: The analysis time.

numberOfSubjects: The number of subjects under consideration when the (interim) analysis takes place.

eventsPerStage1: The observed number of events per stage in treatment group 1.

eventsPerStage2: The observed number of events per stage in treatment group 2.

eventsPerStage: The observed number of events per stage in both treatment groups.

rejectPerStage: 1 if null hypothesis can be rejected, 0 otherwise.

eventsNotAchieved: 1 if number of events could not be reached with observed number of subjects, 0 otherwise.

futilityPerStage: 1 if study should be stopped for futility, 0 otherwise.

testStatistic: The test statistic that is used for the test decision, depends on which design was chosen (group sequential, inverse normal, or Fisher combination test)'

logRankStatistic: Z-score statistic which corresponds to a one-sided log-rank test at considered stage.

conditionalPowerAchieved: The conditional power for the subsequent stage of the trial for selected sample size and effect. The effect is either estimated from the data or can be user defined with thetaH1 or pi1H1 and pi2H1.

trialStop: TRUE if study should be stopped for efficacy or futility or final stage, FALSE otherwise.

hazardRatioEstimateLR: The estimated hazard ratio, derived from the log-rank statistic.

A subset of variables is provided for getSimulationMeans(), getSimulationRates(), getSimulationMultiArmMeans(),
getSimulationMultiArmRates(), or getSimulationMultiArmSurvival().

Examples

results <- getSimulationSurvival(
    pi1 = seq(0.3, 0.6, 0.1), pi2 = 0.3, eventTime = 12,
    accrualTime = 24, plannedEvents = 40, maxNumberOfSubjects = 200,
    maxNumberOfIterations = 50
)
data <- getData(results)
head(data)
#>   iterationNumber stageNumber pi1 pi2 hazardRatio analysisTime numberOfSubjects
#> 1               1           1 0.3 0.3           1     20.39452              169
#> 2               2           1 0.3 0.3           1     19.75876              164
#> 3               3           1 0.3 0.3           1     20.04497              167
#> 4               4           1 0.3 0.3           1     18.74970              156
#> 5               5           1 0.3 0.3           1     21.91084              182
#> 6               6           1 0.3 0.3           1     20.20995              168
#>   overallEvents1 overallEvents2 eventsPerStage rejectPerStage eventsNotAchieved
#> 1             17             23             40              0                 0
#> 2             23             17             40              0                 0
#> 3             20             20             40              0                 0
#> 4             26             14             40              1                 0
#> 5             18             22             40              0                 0
#> 6             20             20             40              0                 0
#>   futilityPerStage testStatistic logRankStatistic conditionalPowerAchieved
#> 1                0   -1.02189184      -1.02189184                       NA
#> 2                0    1.17915109       1.17915109                       NA
#> 3                0   -0.30466915      -0.30466915                       NA
#> 4                0    2.30950492       2.30950492                       NA
#> 5                0   -0.71125579      -0.71125579                       NA
#> 6                0    0.03315397       0.03315397                       NA
#>   pValuesSeparate trialStop hazardRatioEstimateLR
#> 1      0.84658395      TRUE             0.7238649
#> 2      0.11916901      TRUE             1.4519106
#> 3      0.61969092      TRUE             0.9081508
#> 4      0.01045779      TRUE             2.0757646
#> 5      0.76153713      TRUE             0.7985810
#> 6      0.48677590      TRUE             1.0105394
dim(data)
#> [1] 200  19