Parameter Description: Show Source

Parameter Description: Show Source

Arguments

showSource

Logical. If TRUE, the parameter names of the object will be printed which were used to create the plot; that may be, e.g., useful to check the values or to create own plots with the base R plot function. Alternatively showSource can be defined as one of the following character values:

"commands": returns a character vector with plot commands

"axes": returns a list with the axes definitions

"test": all plot commands will be validated with eval(parse()) and returned as character vector (function does not stop if an error occurs)

"validate": all plot commands will be validated with eval(parse()) and returned as character vector (function stops if an error occurs)

Note: no plot object will be returned if showSource is a character.