Class for Simulation Results Survival

A class for simulation results survival.

Details

Use getSimulationSurvival() to create an object of this type.

SimulationResultsSurvival is the basic class for

SimulationResultsSurvival,

SimulationResultsMultiArmSurvival, and

SimulationResultsEnrichmentSurvival.

Fields

maxNumberOfIterations

The number of simulation iterations. Is a numeric vector of length 1 containing a whole number.

seed

The seed used for random number generation. Is a numeric vector of length 1.

allocationRatioPlanned

The planned allocation ratio (n1 / n2) for the groups. For multi-arm designs, it is the allocation ratio relating the active arm(s) to the control. Is a positive numeric vector of length 1.

conditionalPower

The conditional power at each stage of the trial. Is a numeric vector of length 1 containing a value between 0 and 1.

iterations

The number of iterations used for simulations. Is a numeric vector of length 1 containing a whole number.

futilityPerStage

The per-stage probabilities of stopping the trial for futility. Is a numeric matrix.

futilityStop

In simulation results data set: indicates whether trial is stopped for futility or not.

directionUpper

Specifies the direction of the alternative, only applicable for one-sided testing. Default is TRUE which means that larger values of the test statistics yield smaller p-values. Is a logical vector of length 1.

plannedEvents

Determines the number of cumulated (overall) events in survival designs when the interim stages are planned. For two treatment arms, is the number of events for both treatment arms. For multi-arm designs, refers to the overall number of events for the selected arms plus control. Is a numeric vector of length kMax containing whole numbers.

minNumberOfEventsPerStage

Determines the minimum number of events per stage for data-driven sample size recalculation. Is a numeric vector of length kMax containing whole numbers.

maxNumberOfEventsPerStage

Determines the maximum number of events per stage for data-driven sample size recalculation. Is a numeric vector of length kMax containing whole numbers.

thetaH1

The assumed effect under the alternative hypothesis. For survival designs, refers to the hazard ratio. Is a numeric vector.

calcEventsFunction

An optional function that can be entered to define how event size is recalculated. By default, recalculation is performed with conditional power with specified minNumberOfEventsPerStage and maxNumberOfEventsPerStage.

expectedNumberOfEvents

The expected number of events under specified alternative. Is a numeric vector.

pi1

The assumed event rate in the treatment group. Is a numeric vector of length kMax containing values between 0 and 1.

pi2

The assumed event rate in the control group. Is a numeric vector of length 1 containing a value between 0 and 1.

median1

The assumed median survival time in the treatment group. Is a numeric vector.

median2

The assumed median survival time in the reference group. Is a numeric vector of length 1.

maxNumberOfSubjects

The maximum number of subjects for power calculations. Is a numeric vector.

accrualTime

The assumed accrual time intervals for the study. Is a numeric vector.

accrualIntensity

The absolute accrual intensities. Is a numeric vector of length kMax.

dropoutRate1

The assumed drop-out rate in the treatment group. Is a numeric vector of length 1 containing a value between 0 and 1.

dropoutRate2

The assumed drop-out rate in the control group. Is a numeric vector of length 1 containing a value between 0 and 1.

dropoutTime

The assumed time for drop-out rates in the control and treatment group. Is a numeric vector of length 1.

eventTime

The assumed time under which the event rates are calculated. Is a numeric vector of length 1.

thetaH0

The difference or assumed effect under H0. Is a numeric vector of length 1.

allocation1

The number of subjects to be assigned to treatment 1 in subsequent order. Is a numeric vector of length 1 containing a whole number.

allocation2

The number of subjects to be assigned to treatment 2 in subsequent order. Is a numeric vector of length 1 containing a whole number.

kappa

The shape of the Weibull distribution if kappa!=1. Is a numeric vector of length 1.

piecewiseSurvivalTime

The time intervals for the piecewise definition of the exponential survival time cumulative distribution function. Is a numeric vector.

lambda1

The assumed hazard rate in the treatment group. Is a numeric vector of length kMax.

lambda2

The assumed hazard rate in the reference group. Is a numeric vector of length 1.

earlyStop

The probability to stopping the trial either for efficacy or futility. Is a numeric vector.

hazardRatio

The hazard ratios under consideration. Is a numeric vector of length kMax.

studyDuration

The study duration for specified effect size. Is a positive numeric vector.

eventsNotAchieved

The simulated number of cases how often the number of events was not reached. Is a numeric matrix.

numberOfSubjects

In simulation results data set: The number of subjects under consideration when the interim analysis takes place.

numberOfSubjects1

In simulation results data set: The number of subjects under consideration in treatment arm 1 when the interim analysis takes place.

numberOfSubjects2

In simulation results data set: The number of subjects under consideration in treatment arm 2 when the interim analysis takes place.

singleEventsPerStage

The single number of events per stage. Is a numeric matrix.

cumulativeEventsPerStage

The cumulative number of events per stage. Is a numeric matrix.

expectedNumberOfSubjects

The expected number of subjects under specified alternative.

rejectPerStage

The probability to reject a hypothesis per stage of the trial. Is a numeric matrix.

overallReject

The overall rejection probability. Is a numeric vector.

conditionalPowerAchieved

The calculated conditional power, under the assumption of observed or assumed effect sizes. Is a numeric matrix.