Get Design Characteristics

Calculates the characteristics of a design and returns it.

Usage

getDesignCharacteristics(design = NULL, ...)

Arguments

design

The trial design.

...

Ensures that all arguments (starting from the "...") are to be named and that a warning will be displayed if unknown arguments are passed.

Value

Returns a TrialDesignCharacteristics object. The following generics (R generic functions) are available for this result object:

names() to obtain the field names,

print() to print the object,

summary() to display a summary of the object,

plot() to plot the object,

as.data.frame() to coerce the object to a data.frame,

as.matrix() to coerce the object to a matrix.

Details

Calculates the inflation factor (IF), the expected reduction in sample size under H1, under H0, and under a value in between H0 and H1. Furthermore, absolute information values are calculated under the prototype case testing H0: mu = 0 against H1: mu = 1.

How to get help for generic functions

Click on the link of a generic in the list above to go directly to the help documentation of the rpact specific implementation of the generic. Note that you can use the R function methods to get all the methods of a generic and to identify the object specific name of it, e.g., use methods("plot") to get all the methods for the plot generic. There you can find, e.g., plot.AnalysisResults and obtain the specific help documentation linked above by typing ?plot.AnalysisResults.

See also

Examples

# Calculate design characteristics for a three-stage O'Brien & Fleming 
# design at power 90% and compare it with Pocock's design.  
getDesignCharacteristics(getDesignGroupSequential(beta = 0.1))
#> Design parameters and output of group sequential design:
#> 
#> User defined parameters:
#>   Information rates                            : 0.333, 0.667, 1.000 
#>   Type II error rate                           : 0.1000 
#> 
#> Derived from user defined parameters: not available
#> 
#> Default parameters:
#>   Type of design                               : O'Brien & Fleming 
#>   Maximum number of stages                     : 3 
#>   Stages                                       : 1, 2, 3 
#>   Significance level                           : 0.0250 
#>   Two-sided power                              : FALSE 
#>   Test                                         : one-sided 
#>   Tolerance                                    : 1e-08 
#> 
#> Output:
#>   Cumulative alpha spending                    : 0.0002592, 0.0071601, 0.0250000 
#>   Critical values                              : 3.471, 2.454, 2.004 
#>   Stage levels (one-sided)                     : 0.0002592, 0.0070554, 0.0225331 
#> 
#> Group sequential design characteristics:
#>   Number of subjects fixed                     : 10.5074 
#>   Shift                                        : 10.6766 
#>   Inflation factor                             : 1.0161 
#>   Informations                                 : 3.559, 7.118, 10.677 
#>   Power                                        : 0.05653, 0.58531, 0.90000 
#>   Rejection probabilities under H1             : 0.05653, 0.52878, 0.31469 
#>   Futility probabilities under H1              : 0, 0 
#>   Ratio expected vs fixed sample size under H1 : 0.7987 
#>   Ratio expected vs fixed sample size under a value between H0 and H1 : 0.9695 
#>   Ratio expected vs fixed sample size under H0 : 1.0136 
#> 
getDesignCharacteristics(getDesignGroupSequential(beta = 0.1, typeOfDesign = "P")) 
#> Design parameters and output of group sequential design:
#> 
#> User defined parameters:
#>   Type of design                               : Pocock 
#>   Information rates                            : 0.333, 0.667, 1.000 
#>   Type II error rate                           : 0.1000 
#> 
#> Derived from user defined parameters: not available
#> 
#> Default parameters:
#>   Maximum number of stages                     : 3 
#>   Stages                                       : 1, 2, 3 
#>   Significance level                           : 0.0250 
#>   Two-sided power                              : FALSE 
#>   Test                                         : one-sided 
#>   Tolerance                                    : 1e-08 
#> 
#> Output:
#>   Cumulative alpha spending                    : 0.01103, 0.01897, 0.02500 
#>   Critical values                              : 2.289, 2.289, 2.289 
#>   Stage levels (one-sided)                     : 0.01103, 0.01103, 0.01103 
#> 
#> Group sequential design characteristics:
#>   Number of subjects fixed                     : 10.5074 
#>   Shift                                        : 12.0902 
#>   Inflation factor                             : 1.1506 
#>   Informations                                 : 4.030, 8.060, 12.090 
#>   Power                                        : 0.3890, 0.7311, 0.9000 
#>   Rejection probabilities under H1             : 0.3890, 0.3421, 0.1689 
#>   Futility probabilities under H1              : 0, 0 
#>   Ratio expected vs fixed sample size under H1 : 0.7210 
#>   Ratio expected vs fixed sample size under a value between H0 and H1 : 1.0269 
#>   Ratio expected vs fixed sample size under H0 : 1.1391 
#>